Reorganized project directories | Implemented a Makefile | Implemented memory segmentation and segment registers | Fixed a bug inside the assembler

+++ src/cpu.h
@@ -6,7 +6,7 @@
 #define CYCLES_PER_SECOND 10000000  // 10 MHz
 
 void init_cpu(CPU *cpu);
-uint16_t load_program(CPU *cpu, uint8_t *program);
+uint32_t load_program(CPU *cpu, uint8_t *program);
 int step_program(CPU *cpu);
 void cpu_interrupt(CPU *cpu, uint16_t status, Instruction inst);
 void cpu_exception(CPU *cpu, uint16_t status, Instruction inst);
